requirements.txt
anyio==4.9.0
certifi==2025.1.31
greenlet==3.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
python-telegram-bot==20.3
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.40
Telethon==1.39.0
typing_extensions==4.13.2
